avantika/backend/requirements.txt
django>=4.2,<5.0
djangorestframework>=3.14
django-cors-headers>=4.3
python-decouple>=3.8
gunicorn>=21.2
psycopg2-binary>=2.9
numpy>=1.26
scikit-learn>=1.4
